% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxphf-tidiers.R
\name{augment.coxphf}
\alias{augment.coxphf}
\title{Augment a coxphf object}
\usage{
augment.coxphf(x, data = x$y, ...)
}
\arguments{
\item{x}{A `coxphf` object.}

\item{data}{the dataset used to fit the model.}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A tidy [tibble::tibble()] summarizing component-level
  information about the model
}
\description{
Augment a coxphf object
}
