% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxphf-tidiers.R
\name{glance.coxphf}
\alias{glance.coxphf}
\title{Glance at a coxphf object}
\usage{
\method{glance}{coxphf}(x, ...)
}
\arguments{
\item{x}{A `coxphf` object.}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A tidy [tibble::tibble()] summarizing component-level
  information about the model
}
\description{
Glance at a coxphf object
}
